+++ ChangeLog.md
@@ -6,7 +6,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Changed
 
-* `secondsToTimeString()` returns strings in a more sensible format
+* Improved formatting of `Timer::secondsToTimeString()` result
 
 ## [3.0.0] - 2020-02-07