+++ ChangeLog.md
@@ -6,6 +6,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Changed
 
+* `Timer::start()` and `Timer::stop()` now use `hrtime()` (high resolution monotonic timer) instead of `microtime()`
 * Improved formatting of `Timer::secondsToTimeString()` result
 
 ## [3.0.0] - 2020-02-07