+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 All notable changes are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [3.1.1] - 2020-04-17
+
+### Fixed
+
+* [#30](https://github.com/sebastianbergmann/php-timer/pull/30): Resolution of time returned by `Timer::stop()` is different than before (this reverts using `hrtime()` instead of `microtime()`)
+
 ## [3.1.0] - 2020-04-17
 
 ### Added
@@ -48,6 +54,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 * This component is no longer supported on PHP 5.3, PHP 5.4, PHP 5.5, PHP 5.6, and PHP 7.0
 
+[3.1.1]: https://github.com/sebastianbergmann/diff/compare/3.1.0...3.1.1
 [3.1.0]: https://github.com/sebastianbergmann/diff/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/sebastianbergmann/diff/compare/2.1.2...3.0.0
 [2.1.2]: https://github.com/sebastianbergmann/diff/compare/2.1.1...2.1.2